Refactoring getAndPublishCPUMemory
